rename sharding resources to sql resources
