+++ shardingsphere-sql-parser/shardingsphere-sql-parser-test/src/test/java/org/apache/shardingsphere/sql/parser/integrate/asserts/SQLStatementAssert.java
@@ -27,7 +27,7 @@ import org.apache.shardingsphere.sql.parser.integrate.asserts.predicate.Predicat
 import org.apache.shardingsphere.sql.parser.integrate.asserts.selectitem.SelectItemAssert;
 import org.apache.shardingsphere.sql.parser.integrate.asserts.table.AlterTableAssert;
 import org.apache.shardingsphere.sql.parser.integrate.asserts.table.TableAssert;
-import org.apache.shardingsphere.sql.parser.integrate.jaxb.ShardingParserResultSetRegistry;
+import org.apache.shardingsphere.sql.parser.integrate.jaxb.ParserResultSetRegistryFactory;
 import org.apache.shardingsphere.sql.parser.integrate.jaxb.root.ParserResult;
 import org.apache.shardingsphere.sql.parser.sql.segment.dml.item.SelectItemsSegment;
 import org.apache.shardingsphere.sql.parser.sql.segment.dml.order.GroupBySegment;
@@ -80,9 +80,9 @@ public final class SQLStatementAssert {
     
     public SQLStatementAssert(final SQLStatement actual, final String sqlCaseId, final SQLCaseType sqlCaseType, final String databaseType) {
         SQLStatementAssertMessage assertMessage = new SQLStatementAssertMessage(
-                SQLCasesRegistry.getInstance().getSqlCasesLoader(), ShardingParserResultSetRegistry.getInstance().getRegistry(), sqlCaseId, sqlCaseType);
+                SQLCasesRegistry.getInstance().getSqlCasesLoader(), ParserResultSetRegistryFactory.getInstance().getRegistry(), sqlCaseId, sqlCaseType);
         this.actual = actual;
-        expected = ShardingParserResultSetRegistry.getInstance().getRegistry().get(sqlCaseId);
+        expected = ParserResultSetRegistryFactory.getInstance().getRegistry().get(sqlCaseId);
         tableAssert = new TableAssert(assertMessage);
         indexAssert = new IndexAssert(sqlCaseType, assertMessage);
         groupByAssert = new GroupByAssert(assertMessage);
