+++ shardingsphere-sql-parser/shardingsphere-sql-parser-test/src/test/resources/sql/dcl/alter_role.xml
