+++ shardingsphere-sql-parser/shardingsphere-sql-parser-test/src/test/resources/sql/dcl/drop_user.xml
