+++ shardingsphere-sql-parser/shardingsphere-sql-parser-test/src/test/resources/sql/dcl/grant_user.xml
