+++ shardingsphere-sql-parser/shardingsphere-sql-parser-test/src/test/resources/sql/dcl/rename_user.xml
