+++ shardingsphere-sql-parser/shardingsphere-sql-parser-test/src/test/resources/sql/dcl/set_role.xml
