+++ shardingsphere-sql-parser/shardingsphere-sql-parser-test/src/test/resources/sql/ddl/alter.xml
