+++ shardingsphere-sql-parser/shardingsphere-sql-parser-test/src/test/resources/sql/dml/select_aggregate.xml
