+++ shardingsphere-sql-parser/shardingsphere-sql-parser-test/src/test/resources/sql/dml/select_group_by.xml
